Disabled docker login

+++ device-setup.sh
@@ -7,39 +7,48 @@ fi
 
 #Get the registry and application details from the user.
 
-read -p 'Login URL: ' loginurl
-read -p 'Application image name($loginurl/appname)' : APP_NAME
+read -p 'Login URL: ' LOGIN_URL
+read -p 'Image name $(LOGIN_URL)/image: ' APP_NAME
 
-IMAGE= "$loginurl/$APP_NAME"
+IMAGE="$LOGIN_URL/$APP_NAME"
 
 #Docker login into registry 
-docker login $loginurl
-
-#Run the image
-docker run $IMAGE
+#TODO:  check if you need to login
+#docker login $LOGIN_URL
 
 watch_for_updates()
 {
+  echo Pulling Latest $IMAGE
   docker pull $IMAGE
-  CID = $(docker ps | grep $IMAGE| awk '{print $1}') 
+  CID=$(docker ps | grep $IMAGE| awk '{print $1}') 
 
-  for im in CID
+  if [ -z "$CID" ]; then
+     echo "No running container so running and returning"
+     docker run -d $IMAGE
+     return;
+  fi
+
+  for im in $CID
   do
-    LATEST=`docker insepct --format "{{.Id}}" $IMAGE`
+    LATEST=`docker inspect --format "{{.Id}}" $IMAGE`
     RUNNING=`docker inspect --format "{{.Image}}" $im`
     NAME=`docker inspect --format '{{.Name}}' $im | sed "s/\///g"`
     echo "Latest : " $LATEST 
     echo "Running: " $RUNNING
     if [ "$RUNNING" != "$LATEST" ];then
         echo "upgrading $NAME"
-        stop docker-$NAME
+        docker stop $NAME
         docker rm -f $NAME
-        start docker-$NAME
+        docker run -d $NAME
     fi
   done
 }
 
 
 #Every 10 seconds check for updates 
-forever 10 watch_for_updates
+while true
+do
+   watch_for_updates
+   sleep 10
+done
